Return all the alerts in cloudwatch (#102)
* Return all the alerts in cloudwatch

The initial dash implementation didn't take into account that Amazon only
returns a set number of checks per request. We should be able to return
all the cloudwatch checks in Amazon. In order to achieve that, we need
to keep track of the next token before creating a new request and then
stop when the next token is null.

* Only one AmazonCloudWatch client to be created per check

This is a small refactor so that we don't create more than one
AmazonCloudWatch client per check.